amelioration de la fonction delete du backend

+++ controllers/stuff.js
@@ -1,4 +1,5 @@
 const Thing = require('../models/Thing')
+const fs = require('fs')
 
 exports.createThing = (req , res, next)=>{
     
@@ -39,11 +40,24 @@ exports.modifyThing = (req, res, next) => {
        });
 };
 
-exports.deleteThing = (req , res , next) =>{
-    Thing.deleteOne({ _id : req.params.id})
-    .then(() => res.status(200).json({message : 'objet supprimé'}))
-    .catch(error => req.status(400).json({error}))
-};
+exports.deleteThing = (req, res, next) => {
+    Thing.findOne({ _id: req.params.id})
+        .then(thing => {
+            if (thing.userId != req.auth.userId) {
+                res.status(401).json({message: 'Not authorized'});
+            } else {
+                const filename = thing.imageUrl.split('/images/')[1];
+                fs.unlink(`images/${filename}`, () => {
+                    Thing.deleteOne({_id: req.params.id})
+                        .then(() => { res.status(200).json({message: 'Objet supprimé !'})})
+                        .catch(error => res.status(401).json({ error }));
+                });
+            }
+        })
+        .catch( error => {
+            res.status(500).json({ error });
+        });
+ };
 
 exports.getOneThing = ( req , res , next) => {
     Thing.findOne({_id : req.params.id})
